Improve parameterising of selectable hashing algorithm
