+++ src/crypto/CMakeLists.txt
@@ -45,9 +45,9 @@ set(crypto_sources
   oaes_lib.c
   random.c
   skein.c
-  cn_slow_hash_hard_arm.cpp
-  cn_slow_hash_hard_intel.cpp
-  cn_slow_hash_soft.cpp
+  cn_heavy_hash_hard_arm.cpp
+  cn_heavy_hash_hard_intel.cpp
+  cn_heavy_hash_soft.cpp
   tree-hash.c)
 
 set(crypto_headers)
@@ -70,7 +70,7 @@ set(crypto_private_headers
   random.h
   skein.h
   skein_port.h
-  cn_slow_hash.hpp)
+  cn_heavy_hash.hpp)
 
 loki_private_headers(cncrypto
   ${crypto_private_headers})
