+++ src/crypto/hash.h
@@ -38,7 +38,7 @@
 #include "generic-ops.h"
 #include "hex.h"
 #include "span.h"
-#include "crypto/cn_slow_hash.hpp"
+#include "crypto/cn_heavy_hash.hpp"
 
 namespace crypto {
 
@@ -72,24 +72,18 @@ namespace crypto {
     return h;
   }
 
-  inline void cn_slow_hash(const void *data, std::size_t length, hash &hash, int variant = 0) {
-    static thread_local cn_pow_hash_v2 ctx2;
-    static thread_local cn_pow_hash_v1 ctx1 = cn_pow_hash_v1::make_borrowed(ctx2);
-    if (variant == 0) {
-      ctx1.hash(data, length, hash.data);
-    } else {
-      ctx2.hash(data, length, hash.data);
-    }
-  }
+  enum struct cn_slow_hash_type
+  {
+      heavy_v1,
+      heavy_v2,
+  };
 
-  inline void cn_slow_hash_prehashed(const void *data, std::size_t length, hash &hash, int variant = 0) {
-    static thread_local cn_pow_hash_v2 ctx2;
-    static thread_local cn_pow_hash_v1 ctx1 = cn_pow_hash_v1::make_borrowed(ctx2);
-    if (variant == 0) {
-      ctx1.hash(data, length, hash.data, true);
-    } else {
-      ctx2.hash(data, length, hash.data, true);
-    }
+  inline void cn_slow_hash(const void *data, std::size_t length, hash &hash, cn_slow_hash_type type) {
+    static thread_local cn_heavy_hash_v2 v2;
+    static thread_local cn_heavy_hash_v1 v1 = cn_heavy_hash_v1::make_borrowed(v2);
+
+    if (type == cn_slow_hash_type::heavy_v1) v1.hash(data, length, hash.data);
+    else                                     v2.hash(data, length, hash.data);
   }
 
   inline void tree_hash(const hash *hashes, std::size_t count, hash &root_hash) {
