+++ tests/performance_tests/cn_slow_hash.h
@@ -62,7 +62,7 @@ public:
   bool test()
   {
     crypto::hash hash;
-    crypto::cn_slow_hash(&m_data, sizeof(m_data), hash);
+    crypto::cn_slow_hash(&m_data, sizeof(m_data), hash, crypto::cn_heavy_hash_v1);
     return hash == m_expected_hash;
   }
